Added movement to GenericEntity
GenericEntity will move left and right on it's own. This can be used to prove that Spatial Partitioning is working. When Spatial Partitioning is turned on, there will be a single object on the map moving. However, when it turned off, the previous single object should move differently from others.